refactor: move the AttestTracker type to a new file
