+++ validator/dispatcher.go
@@ -1,149 +1,17 @@
 package validator
 
 import (
-	"errors"
-	"fmt"
 	"strings"
 	"time"
 
-	"github.com/NethermindEth/juno/core/felt"
 	junoUtils "github.com/NethermindEth/juno/utils"
-	"github.com/NethermindEth/starknet-staking-v2/validator/constants"
 	"github.com/NethermindEth/starknet-staking-v2/validator/metrics"
 	signerP "github.com/NethermindEth/starknet-staking-v2/validator/signer"
 	"github.com/NethermindEth/starknet-staking-v2/validator/types"
-	"github.com/NethermindEth/starknet.go/rpc"
-	"github.com/NethermindEth/starknet.go/utils"
 )
 
 // Created a function variable for mocking purposes in tests
 var Sleep = time.Sleep
-
-var ErrTxnHashNotFound = rpc.ErrHashNotFound
-
-type AttestStatus uint8
-
-const (
-	Iddle AttestStatus = iota
-	Ongoing
-	Successful
-	Failed
-)
-
-type AttestTransaction struct {
-	txn   rpc.BroadcastInvokeTxnV3
-	valid bool
-}
-
-func (t *AttestTransaction) Build(signer signerP.Signer, blockHash *types.BlockHash) error {
-	t.valid = false
-
-	var err error
-	t.txn, err = signer.BuildAttestTransaction(blockHash)
-	if err != nil {
-		return fmt.Errorf("signer failed building the transaction: %w", err)
-	}
-
-	_, err = signer.SignTransaction(&t.txn)
-	if err != nil {
-		return fmt.Errorf("signer failed to sign the transaction: %w", err)
-	}
-
-	t.valid = true
-
-	return nil
-}
-
-func (t *AttestTransaction) Invoke(signer signerP.Signer) (
-	rpc.AddInvokeTransactionResponse, error,
-) {
-	var resp rpc.AddInvokeTransactionResponse
-	if !t.valid {
-		return resp, errors.New("invoking attest transaction before building it")
-	}
-	t.valid = false
-
-	// todo(rdr): make sure to estimate fee with query bit with Braavos Account
-	estimate, err := signer.EstimateFee(&t.txn)
-	if err != nil {
-		return resp, fmt.Errorf("signer failed to estimate fee: %w", err)
-	}
-	t.txn.ResourceBounds = utils.FeeEstToResBoundsMap(estimate, constants.FeeEstimationMultiplier)
-
-	// patch for making sure txn.Version is correct
-	t.txn.Version = rpc.TransactionV3
-
-	_, err = signer.SignTransaction(&t.txn)
-	if err != nil {
-		return resp, fmt.Errorf("signer failed to sign the transaction: %w", err)
-	}
-
-	resp, err = signer.InvokeTransaction(&t.txn)
-	if err != nil {
-		return resp, fmt.Errorf("signer failed to invoke the transaction: %w", err)
-	}
-
-	return resp, nil
-}
-
-func (t *AttestTransaction) UpdateNonce(signer signerP.Signer) error {
-	if !t.valid {
-		return errors.New("updating the transaction nonce before building the transaction")
-	}
-	newNonce, err := signer.Nonce()
-	if err != nil {
-		return fmt.Errorf("signer failed to get the nonce: %w", err)
-	}
-	if !t.txn.Nonce.Equal(newNonce) {
-		t.txn.Nonce = newNonce
-		_, err := signer.SignTransaction(&t.txn)
-		if err != nil {
-			return fmt.Errorf("signer failed to sign the transaction: %w", err)
-		}
-	}
-
-	return nil
-}
-
-// I want to name this built or smth like that
-func (t *AttestTransaction) Valid() bool {
-	return t.valid
-}
-
-type AttestTracker struct {
-	Transaction AttestTransaction
-	Hash        felt.Felt
-	Status      AttestStatus
-}
-
-func NewAttestTracker() AttestTracker {
-	//nolint:exhaustruct // Using default values
-	return AttestTracker{
-		Transaction: AttestTransaction{},
-		Status:      Iddle,
-	}
-}
-
-func (a *AttestTracker) UpdateStatus(
-	signer signerP.Signer,
-	logger *junoUtils.ZapLogger,
-) {
-	status := TrackAttest(signer, logger, &a.Hash)
-	a.setStatus(status)
-}
-
-func (a *AttestTracker) setStatus(status AttestStatus) {
-	a.Status = status
-	switch status {
-	case Ongoing, Successful:
-	case Failed:
-		a.Hash = felt.Zero
-	case Iddle:
-		panic("status cannot be change to iddle")
-	default:
-		panic(fmt.Sprintf("unknown tracker status %d", status))
-	}
-}
 
 type EventDispatcher[S signerP.Signer] struct {
 	// Event channels
@@ -208,7 +76,7 @@ func (d *EventDispatcher[S]) Dispatch(
 					continue
 				}
 			}
-			d.CurrentAttest.setStatus(Ongoing)
+			d.CurrentAttest.SetStatus(Ongoing)
 
 			// Case when the validator is initiated mid window and didn't have time to prepare
 			// or the transaction invoke failed.
@@ -244,7 +112,7 @@ func (d *EventDispatcher[S]) Dispatch(
 					logger.Infow(
 						"attestation is already done for this epoch",
 					)
-					d.CurrentAttest.setStatus(Successful)
+					d.CurrentAttest.SetStatus(Successful)
 
 					continue
 				}
@@ -253,7 +121,7 @@ func (d *EventDispatcher[S]) Dispatch(
 					"failed to attest",
 					"error", err.Error(),
 				)
-				d.CurrentAttest.setStatus(Failed)
+				d.CurrentAttest.SetStatus(Failed)
 
 				continue
 			}
@@ -287,60 +155,5 @@ func (d *EventDispatcher[S]) Dispatch(
 			// check the account balance
 			go CheckBalance(signer, balanceThreshold, logger, tracer)
 		}
-	}
-}
-
-func TrackAttest[S signerP.Signer](
-	signer S,
-	logger *junoUtils.ZapLogger,
-	txHash *felt.Felt,
-) AttestStatus {
-	txStatus, err := signer.TransactionStatus(txHash)
-	if err != nil {
-		if err.Error() == ErrTxnHashNotFound.Error() {
-			logger.Infow(
-				"attest transaction status was not found. Will wait.",
-				"transaction hash", txHash,
-			)
-
-			return Ongoing
-		} else {
-			logger.Errorw(
-				"attest transaction FAILED. Will retry.",
-				"transaction hash", txHash,
-				"error", err,
-			)
-
-			return Failed
-		}
-	}
-
-	if txStatus.ExecutionStatus == rpc.TxnExecutionStatusREVERTED {
-		logger.Errorw(
-			"attest transaction REVERTED. Will retry.",
-			"transaction hash", txHash,
-			"failure reason", txStatus.FailureReason,
-		)
-
-		return Failed
-	}
-
-	switch txStatus.FinalityStatus {
-	case rpc.TxnStatusReceived, rpc.TxnStatusCandidate, rpc.TxnStatusPreConfirmed:
-		logger.Infow(
-			fmt.Sprintf("attest transaction %s. Will wait.", txStatus.FinalityStatus),
-			"hash", txHash,
-		)
-
-		return Ongoing
-	default:
-		logger.Infow(
-			"attest transaction SUCCESSFUL.",
-			"transaction hash", txHash,
-			"finality status", txStatus.FinalityStatus,
-			"execution status", txStatus.ExecutionStatus,
-		)
-
-		return Successful
 	}
 }
